requirements.txt
streamlit==1.28.1
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
pandas==2.1.3
plotly==5.17.0
supabase==2.0.2
requests==2.31.0 